Update README.md
Jokes.py had improvements to answers that were coded in earlier versions of Python. Input() did not work in my (Python 3) environment when I first tried the tutorial, so I substituted raw_input, and it worked. I'm not sure what changed in a few weeks, but now input is the only thing that works in Idle.